Rename utils/type.ts to utils/array.ts
